[fix](ut)Fix StmtExecutorTest (#50322)
Fix StmtExecutorTest
